+++ spec/osm/event_spec.rb
@@ -69,7 +69,12 @@ describe "Event" do
         :member_id => 1,
         :grouping_id => 2,
         :row => 3,
-        :columns => [],
+        :first_name => 'First',
+        :last_name => 'Last',
+        :attending => :yes,
+        :date_of_birth => Date.new(2000, 1, 2),
+        :fields => {},
+        :payments => {},
         :event => Osm::Event.new(:id => 1, :section_id => 1, :name => 'Name', :columns => [])
       }
 
@@ -77,7 +82,12 @@ describe "Event" do
       ea.member_id.should == 1
       ea.grouping_id.should == 2
       ea.fields.should == {}
+      ea.payments.should == {}
       ea.row.should == 3
+      ea.first_name.should == 'First'
+      ea.last_name.should == 'Last'
+      ea.date_of_birth.should == Date.new(2000, 1, 2)
+      ea.attending.should == :yes
       ea.valid?.should be_true
     end
 
@@ -528,6 +538,8 @@ describe "Event" do
             'dob' => '1980-01-02',
             'patrolid' => '2',
             'f_1' => 'a',
+            'payment' => 'Manual',
+            'p1' => ''
           }
         ]
       }
@@ -541,25 +553,29 @@ describe "Event" do
       ea = attendance[0]
       ea.member_id.should == 1
       ea.grouping_id.should == 2
+      ea.first_name.should == 'First'
+      ea.last_name.should == 'Last'
+      ea.date_of_birth.should == Date.new(1980, 1, 2)
+      ea.attending.should == :yes
       ea.fields.should == {
-        'firstname' => 'First',
-        'lastname' => 'Last',
-        'dob' => Date.new(1980, 1, 2),
-        'attending' => true,
-        'f_1' => 'a',
+        1 => 'a',
+      }
+      ea.payments.should == {
+        1 => '',
       }
       ea.row.should == 0
     end
 
     it "Update attendance (succeded)" do
-      ea = Osm::Event::Attendance.new(:row => 0, :member_id => 4, :fields => {'f_1' => 'TEST'}, :event => Osm::Event.new(:id => 2, :section_id => 1))
+      ea = Osm::Event::Attendance.new(:row => 0, :member_id => 4, :fields => {1 => 'old value', 2 => 'another old value'}, :event => Osm::Event.new(:id => 2, :section_id => 1))
 
+      ea.fields[1] = 'value'
       HTTParty.should_receive(:post).with(
         "https://www.onlinescoutmanager.co.uk/events.php?action=updateScout",
         {:body => {
           'scoutid' => 4,
           'column' => 'f_1',
-          'value' => 'TEST',
+          'value' => 'value',
           'sectionid' => 1,
           'row' => 0,
           'eventid' => 2,
@@ -570,7 +586,41 @@ describe "Event" do
         }}
       ) { OsmTest::DummyHttpResult.new(:response=>{:code=>'200', :body=>'{}'}) }
 
-      ea.update(@api, 'f_1').should be_true
+      ea.attending = :yes
+      HTTParty.should_receive(:post).with(
+        "https://www.onlinescoutmanager.co.uk/events.php?action=updateScout",
+        {:body => {
+          'scoutid' => 4,
+          'column' => 'attending',
+          'value' => 'Yes',
+          'sectionid' => 1,
+          'row' => 0,
+          'eventid' => 2,
+          'apiid' => @CONFIGURATION[:api][:osm][:id],
+          'token' => @CONFIGURATION[:api][:osm][:token],
+          'userid' => 'user_id',
+          'secret' => 'secret',
+        }}
+      ) { OsmTest::DummyHttpResult.new(:response=>{:code=>'200', :body=>'{}'}) }
+
+      ea.payment_control = :automatic
+      HTTParty.should_receive(:post).with(
+        "https://www.onlinescoutmanager.co.uk/events.php?action=updateScout",
+        {:body => {
+          'scoutid' => 4,
+          'column' => 'payment',
+          'value' => 'Automatic',
+          'sectionid' => 1,
+          'row' => 0,
+          'eventid' => 2,
+          'apiid' => @CONFIGURATION[:api][:osm][:id],
+          'token' => @CONFIGURATION[:api][:osm][:token],
+          'userid' => 'user_id',
+          'secret' => 'secret',
+        }}
+      ) { OsmTest::DummyHttpResult.new(:response=>{:code=>'200', :body=>'{}'}) }
+
+      ea.update(@api).should be_true
     end